#19 - shifted files under internal/cyberark #19 - renamed provider files for clarity #18 - Remove delete feature, updated logging
